WebhookCreator/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
discord.py==2.2.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pip==23.1.2
python-dotenv==1.0.0
sentry-sdk==1.25.1
setuptools==67.8.0
urllib3==1.26.16
yarl==1.9.2
